requirements.txt
streamlit
google-generativeai
python-dotenv
SpeechRecognition==3.10.1
PyAudio==0.2.14
roslibpy==1.6.0